fix(records): extract RepeatableRowsField so create.tsx stays under its length cap
Folding reorder into RepeatableRowsField grew create.tsx past the line-count
waiver the fe-file-length gate froze it at, which fails in both the frontend and
the deterministic-gates lanes (the gate runs in both). The component is
self-contained, so it moves to its own file — a screen file is not where a
growing control belongs — and create.tsx drops back under the frozen count,
which is then tightened to the new length.

No behaviour change: RepeatableRowsField is imported back into create.tsx and
renders identically; fieldControl/fieldLabel are hoisted, so the create ⇄
repeatablerowsfield import cycle resolves at load and again at render.
