+++ scripts/fe-file-length-waivers.txt
@@ -65,7 +65,7 @@
 940 frontend/src/screens/contacts.tsx
 667 frontend/src/screens/contractform.tsx
 528 frontend/src/screens/contractlifecycle.tsx
-982 frontend/src/screens/create.tsx
+882 frontend/src/screens/create.tsx
 938 frontend/src/screens/customfields.tsx
 617 frontend/src/screens/dealroomaccess.tsx
 2633 frontend/src/screens/deals.test.tsx
